Update dependencies and adjust SeaORM features
Updated SeaORM and SeaORM migration features from SQLite to Postgres in Cargo.toml. Bumped versions for minil-config and minil-migration, and added new dependencies. Cleaned up .gitignore and fixed an import alias in inflector.rs.

+++ main/minil/src/main.rs
@@ -617,7 +617,7 @@ async fn list_objects(
                                         .id(owner.id)
                                         .build(),
                                 )
-                                .size(object.size as u64)
+                                .size(object.size)
                                 .build()
                         })
                         .collect(),
@@ -724,7 +724,7 @@ async fn list_objects_v2(
                                             .build()
                                     },
                                 ))
-                                .size(object.size as u64)
+                                .size(object.size)
                                 .build()
                         })
                         .collect(),
